requirements.txt
# Smart Gallery Backend - Requirements
# Tested compatible versions for Python 3.10+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# CORS and Security
starlette==0.36.3

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# ML/AI - Core
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0

# YOLO Object Detection
ultralytics==8.1.0

# Face Recognition - InsightFace
insightface==0.7.3
onnxruntime-gpu==1.17.0

# CLIP for Semantic Search
open-clip-torch==2.24.0

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.25

# Vector Search
faiss-cpu==1.7.4

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.26.0
tqdm==4.66.1

# Image EXIF
piexif==1.1.3
